+++ src/main/core/java/io/github/greyp9/arwo/core/io/command/Command.java
@@ -18,7 +18,7 @@ public abstract class Command {
 
     public abstract Date getFinish();
 
-    public abstract Integer getPID();
+    public abstract Long getPID();
 
     public abstract Integer getExitValue();
