+++ src/main/core/java/io/github/greyp9/arwo/core/io/command/CommandToDo.java
@@ -36,7 +36,7 @@ public class CommandToDo extends Command {
     }
 
     @Override
-    public final Integer getPID() {
+    public final Long getPID() {
         return null;
     }
